+++ android/src/main/java/com/bitmovin/player/reactnative/RNPlayerViewPackage.kt
@@ -30,6 +30,7 @@ class RNPlayerViewPackage : ReactPackage {
             BitmovinCastManagerModule(reactContext),
             BufferModule(reactContext),
             NetworkModule(reactContext),
+            DebugModule(reactContext),
         )
     }